requirements.txt
Django==4.0.5
django-rest-framework==0.1.0
drf-yasg==1.20.0
pika==1.3.0
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.2.0
